In is_leapsecmonth(), check for a month values below 1 and add 12 to them.
The configuration file has leap second months in UTC, so typically June
and December. However, in local time, the month value is already one higher
when a leap second is announced (23:mm UTC) or inserted (00:00 UTC).

This means that we have to subtract one from the current local month value
to get the UTC month.