Weaken precondition for processing leap seconds / DST changes to not insist on correct values for p1, p2, p3, and bit 0, 20, 17, 18.
This allows for some bit errors at the time the leap second or DST change occur,
the event must still be annouced and the date and time must still be correct,
either actual or derived.

For processing the announcement the above parities and bit values must
still be correct.

Remove a now obsolete comment.