Replace the bit detection algorithm by an idea from Udo Klein, with permission.
For a raw pulse log which is 5.7 days long the number of wrongly detected
minutes dropped from ~2000 (depending on the value of minlen) to ~20.

The realfreq parameter should be set to the average value of thigh seen
so that the new bit starts at approximately 50% of the rising pulse.

Link readpin against libm