Determine the value of realfreq dynamically instead of reading a static value from config.txt
Use a exponential filter with a 'weight' of 1/20 to determine realfreq.
The exponential filter to determine the pulse value is updated each time
realfreq is updated.

Introduce a flag newminute which keeps the old value of "t > realfreq * 3/2",
because realfreq is updated each but the first second (and when a new
minute arrives, we have to adjust realfreq with only t/2.