Determine maxone and maxzero dynamically, obsoleting them from config.txt.
Initially maxone is set to 32% (maxzero is always half of maxone).
The lengths of the active bubbles for bit 0 and bit 20 (always 0 and 1
respectively) are tracked using length = length/2 + newlength/2 . From
this, maxone is determined as (length bit of bit 0 + length of bit 20)/2
at the start of a new minute.

The initial lengths of bit 0 and bit 20 are 10% and 20% of realfreq. If
the value of bit 0 or bit 20 is wrong, no adjustment for that bit is made.