Convert all floating point operations to corresponding integer operations.
Fields bit0, bit20, and realfreq get multiplied by 10^6. Note that 1e6 is
a floating point number, so write 1000000 instead.

Drop field a, it can be derived from the realfreq field if required.

Properly size the counter fields.

Move the signal buffer to the end of the structure.

Adjust displaying in dcf77pi.c (counters can now be 6 digits wide) and
readpin.c