Do not stamp the time as OK when the minute is exactly one bit too long.
This could happen because for a minute of 60 bits it is only known if it
is too long when the leap second information has been processed, which is
where dt_res.minute_length is finalized. Use that information too in
stamp_date_time().

While here, remove some superfluous if statements that are always true.

Issue #18