Split mainloop into an analyze and live part.
Instead of a mainloop for dcf77pi and dcf77pi-analyze and a very similar
one for nplpi and nplpi-analyze (later on), make a mainloop for
dcf77pi-analyze and nplpi-analyze and another one for the live versions.

Although not strictly necessary now, it should make integrating
timer_create-demo into mainloop_live() easier.